+++ changelog.md
@@ -20,6 +20,7 @@ Codion Change Log
 - DefaultFileSelectionDialogBuilder now adds selected file extension to filename, related refactoring.
 ### is.codion.framework.domain
 - EntityType.define() and AttributeDefiner.define() renamed as().
+- DerivedAttributeDefinition.sources() renamed attributes().
 ### is.codion.framework.db.local
 - DefaultEntityQueries.update() bug fixed, buggy modified check removed, caller is responsible for providing a modified entity.
 - EntityResultPacker, minor refactoring, transient attributes are never derived.